+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a JWK/JWKS representation.
+- Added `HasPublicKey` trait for signers to provide their public key for verification.
+
 ### Breaking
 
 - Removed the sync traits.